Fix running all jest files using global{Setup/Teardown}
