+++ tests/globalTeardown.ts
@@ -1,0 +1,3 @@
+export default async function teardown() {
+  await globalThis.setup.api.disconnect();
+}
