+++ tests/setup.ts
@@ -1,26 +1,6 @@
-import { ApiPromise, WsProvider, Keyring } from '@polkadot/api';
-import { KeyringPair } from '@polkadot/keyring/types';
 import { ReturnNumber } from '@supercolony/typechain-types';
 import Token from '../types/contracts/psp22_token';
 import { expect } from '@jest/globals';
-// Create a new instance of contract
-const wsProvider = new WsProvider('ws://127.0.0.1:9944');
-// Create a keyring instance
-const keyring = new Keyring({ type: 'sr25519' });
-export async function setupApi(): Promise<{
-  api: ApiPromise;
-  alice: KeyringPair;
-  bob: KeyringPair;
-}> {
-  const api = await ApiPromise.create({ provider: wsProvider });
-  const alice = keyring.addFromUri('//Alice');
-  const bob = keyring.addFromUri('//Bob');
-  return {
-    api,
-    alice,
-    bob,
-  };
-}
 
 export function parseUnits(amount: bigint | number, decimals = 18): bigint {
   return BigInt(amount) * 10n ** BigInt(decimals);
@@ -74,14 +54,12 @@ export async function changeTokenBalances<T>(
       async (account) => (await token.query.balanceOf(account)).value.rawNumber,
     ),
   );
-  console.log(beforeBalances);
   const result = await txThunk();
   const afterBalances = await Promise.all(
     accounts.map(
       async (account) => (await token.query.balanceOf(account)).value.rawNumber,
     ),
   );
-  console.log(afterBalances);
   const changes = afterBalances.map((afterBalance, i) =>
     afterBalance.sub(beforeBalances[i]).toString(),
   );
